feat: rename main script and use overlay webcomponent
